In order to facilitate updating the TorrentGroup
activity after changing the bookmark property without having to reload
the page I've added some stuff

- I've added a success/fail return value to MySoup.pressLink based on
the HTTP status code it gets in response to the action
so HTTP 200 OK returns true, anything else is an error and returns
false.
- I've added to api.torrents.torrents.Group the ability to set the
bookmarked property, which is used to update the Group's field
without having to perform re-load of the entire Group. Unfortunately
this method currently is public because it's changed from
api.torrents.torrents.TorrentGroup but I'm contemplating re-structuring
some of the API classes to use inheritance, so TorrentGroup will
inherit from Group so it can easily get all its fields.
- The TorrentGroup.add/removeBookmark will only change the field if
pressLink went ok, and will return the same value returned
by pressLink so the caller of add/remove can check if it went ok as
well.