Add functionality for the new API endpoint user_recents
This API returns lists of the user's recent uploads and snatches giving the album name, the torrent group id, the image url and a list of the artists. The artists field is kind of being ignored for now as it's structured a bit strangely and I'm not sure how to handle it properly with GSON. For now I'll just show the album name in the recents view. It would be nice to show the artist name as well, but I'm not sure what to do with the strange artists API response.