Make it possible to group and sort an artist's releases and appearances
Creating a Releases class will build up the groupings of the Artist's releases and appearances into two SortedMaps (for releases & appearances) where the keys map to sets of the TorrentGroups sorted in descending chronological order. If two albums share the year they're sorted in reverse string order which seems to be what the site does when comparing my ordering to what shows up on the site.

It'd be nice if the API made this task easier but I think this is the best we can do. Looking up the type of appearance made by an artists in the ExtendedArtists object returned by the API really sucks but the lists won't be too long (<25-50 entries avg I bet) so it shouldn't be too bad overall, it just looks ugly.