Making the handling of grouping the releases simpler
We don't distinguish between appearance & release type (which looks like is what the site does as well). This will also make it simpler to deal with in the app since we'll only need to worry about a single map. Because of this the ARTIST in the ReleaseType enum is just used to tell us to look at the actual release type. I also realize now that I've been commiting directly to develop and not in a feature branch but oh well.

Unless I decide on some final tweaks I'll merge this with master and start integrating it into the app.