RequestsSearch getting the same clean up treatment
Making variables that should be per-instance be so, improving prev/next page handling, cleaning up interface, etc. Also properly read the API artists response for requests search as well. The response itself is pretty weird, I'll have to mention it to Gwindow. Also noticed that Artists in the requests have the id returned as a string, which also seems like a mistake.